+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Add an optional id to the st-toggle-button interface
 * Modify the input example with max and min validations
 * Add minor changes to the radio menu example
+* st-switch: Add missing inputs/outputs and fix bad descriptions
 
 ## 1.3.0 (Abril 4, 2017)